Restructure to use sliders and rename classes
