+++ Assets/_BForBoss/_Core/Scripts/Audio.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 3235889403f1846408b1583d81e3c1f4
-PrefabImporter:
+guid: 5dc8d19ff178c2140a6c5d86f1e6be00
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
