+++ Assets/_BForBoss/_Core/Scripts/Audio/BusVolumeModel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 96c3e70ae6d65ac41941c389b2e119f5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
