+++ Assets/_BForBoss/_Core/Scripts/Audio/IVolumeModel.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace BForBoss
+{
+    public interface IVolumeModel
+    {
+        void SetVolume(float volume);
+        bool GetVolume(out float volume);
+    }
+}
