+++ Assets/_BForBoss/_Core/Scripts/Audio/IVolumeModel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fc21aefb3ac8d734c926d38955448872
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
