+++ Assets/_BForBoss/_Core/Scripts/Audio/VCAVolumeModel.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3623e6ccb4330b44cbeedf060401017e
+guid: fcccf7637b1e3ae42ab10d2841f74a1d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
