+++ Assets/_BForBoss/_Core/Scripts/UserInterface/AudioSettingsViewBehaviour.cs
@@ -1,10 +1,17 @@
 using FMODUnity;
+using Perigon.UserInterface;
 using UnityEngine;
 
 namespace BForBoss
 {
     public class AudioSettingsViewBehaviour : MonoBehaviour
     {
+
+        [SerializeField] private SliderBehaviour _mainVolumeSlider;
+        [SerializeField] private SliderBehaviour _musicVolumeSlider;
+        [SerializeField] private SliderBehaviour _sfxVolumeSlider;
+        
+        
         public void Initialize()
         {
